✨ 스낵스 일부 구현 + JWT 수정중
# <이모지> <제목> 의 형식으로 제목을 아래 공백줄에 작성
# 제목은 50자 이내 / 변경사항이 "무엇"인지 명확히 작성 / 끝에 마침표 금지
✨ 로그인 기능 추가

# 바로 아래 공백은 지우지 마세요 (제목과 본문의 분리를 위함)

################
# 본문(구체적인 내용)을 아랫줄에 작성
# 여러 줄의 메시지를 작성할 땐 "-"로 구분 (한 줄은 72자 이내)
- 회원 정보 확인
- 토큰 발급 및 저장
################
# 꼬릿말(footer)을 아랫줄에 작성 (현재 커밋과 관련된 이슈 번호 추가 등)
# Jira Smart commit 사용 시 Jira 이슈 또한 종료할 수 있음.
# 예) Close #7

################
# ✨ : 새로운 기능 추가
# 🐛 : 버그 수정
# 📚 : 문서 수정
# 🚨 : 테스트 코드 추가
# 🔨 : 코드 리팩토
# 📝 : 코드 의미에 영향을 주지 않는 변경사항
# 🔧 : 기타 변경사항, 프로`덕션 코드 변경사항 없음
################

+++ backend-java/src/main/java/com/ssafy/db/repository/SnacksRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 public interface SnacksRepository extends JpaRepository<Snacks, Long> {
 //    Page<Snacks> findAll(Pageable pageable);
 
-    Optional<Snacks> findBySnacksId(int snacksId);
+    Optional<Snacks> findBySnacksId(Long snacksId);
 
     List<Snacks> findSnacksByUserUserIdOrderBySnacksIdDesc(String userId);
